image-recommender/requirements.txt
Flask
mysql-connector-python
numpy
scikit-learn
redis
